forgot to include this stuff

+++ load_app/tin/operations/export_substances.py
@@ -1,15 +1,22 @@
 from load_app.common.consts import *
 from load_app.common.database import Database
-from load_app.tin.common import subid_to_zincid_opt
+from load_app.tin.common import subid_to_zincid_opt,get_tranches
 import os, sys, subprocess, shutil
 
 def export_substances(destination):
 
     export_tmp_dest = BIG_SCRATCH_DIR + "/export_{}_{}".format(Database.instance.host, Database.instance.port)
     export_raw_dest = export_tmp_dest + "/raw_substances"
-    export_raw_split_dest = export_tmp_dest + "/split_substances"
+    export_raw_split_dest = export_tmp_dest + "/split_substances_tranche"
+
+    tranches = get_tranches()
+    first_tranche = tranches[0]
+    first_export = '/'.join([destination, first_tranche[:3], first_tranche+'.smi'])
+    if os.path.exists(first_export) or os.path.exists(first_export + '.gz'):
+        raise Exception("unable to export- remove existing files @ {} ::: {}".format(destination, tranches[0] + '->' + tranches[-1]))
 
     if os.path.exists(export_raw_split_dest):
+        print("removing prev. export")
         shutil.rmtree(export_raw_split_dest)
 
     subprocess.call(["mkdir", "-p", export_raw_split_dest])
